Implementado inativação de categoria

+++ app/Http/Controllers/Categorias.php
@@ -6,6 +6,7 @@ use App\DTO\Categorias\CreateCategorias;
 use App\DTO\Categorias\UpdateCategorias;
 use App\Http\Requests\RequestCategorias;
 use App\Models\Categoria;
+use App\Models\Produto;
 use App\Services\Categorias\CategoriaService;
 use Illuminate\Http\Request;
 
@@ -29,8 +30,10 @@ class Categorias extends Controller
 
         $filters = ['filter' => $request->get('filter', '')];
 
+        $inativas = $this->model->where('ativa', 0)->get();
 
-        return view('dashboard.categorias.show_categorias', compact('categorias', 'filters'));
+
+        return view('dashboard.categorias.show_categorias', compact('categorias', 'filters', 'inativas'));
     }
 
     //=========================================================================================================
@@ -40,7 +43,7 @@ class Categorias extends Controller
 
         return view('dashboard.categorias.detalhes_categorias', compact('categoria'));
     }
-    
+
     //=========================================================================================================
     public function create()
     {
@@ -52,8 +55,12 @@ class Categorias extends Controller
     //=========================================================================================================
     public function store(RequestCategorias $request)
     {
-        if ($this->model->where('categoria', $request->categoria)->first()) {
+        if ($this->model->where('categoria', $request->categoria)->where('ativa', 1)->first()) {
             return redirect()->route('create.categorias')->withInput()->with('error_create', 'Já existe essa categoria registrada.');
+        }
+
+        if ($this->model->where('categoria', $request->categoria)->where('ativa', 0)->first()) {
+            return redirect()->route('create.categorias')->withInput()->with('error_create', 'Já existe essa categoria inativada.');
         }
 
         $this->service->store(
@@ -75,9 +82,12 @@ class Categorias extends Controller
     //=========================================================================================================
     public function update_submit(RequestCategorias $request)
     {
-        if ($this->model->where('id', '!=', $request->id)->where('categoria', $request->categoria)->first())
-        {
+        if ($this->model->where('id', '!=', $request->id)->where('categoria', $request->categoria)->where('ativa', 1)->first()) {
             return redirect()->back()->withInput()->with('error_create', 'Já existe essa categoria.');
+        }
+
+        if ($this->model->where('id', '!=', $request->id)->where('categoria', $request->categoria)->where('ativa', 0)->where('ativa', 0)->first()) {
+            return redirect()->route('create.categorias')->withInput()->with('error_create', 'Já existe essa categoria inativada.');
         }
 
         $this->service->update(
@@ -88,12 +98,49 @@ class Categorias extends Controller
     }
 
     //=========================================================================================================
+    public function inativar_categoria(string $id)
+    {
+        $this->model->findOrFail($id)->update(
+            ['ativa' => 0]
+        );
+
+        return redirect()->route('show.categorias');
+    }
+
+    //=========================================================================================================
+    public function reativar_categoria(string $id)
+    {
+        $this->model->findOrFail($id)->update(
+            ['ativa' => 1]
+        );
+
+        return redirect()->route('show.categorias');
+    }
+
+    //=========================================================================================================
+    public function categorias_inativas(Request $request)
+    {
+        $categorias = $this->service->paginateInativas(
+            page: $request->get('page', 1),
+            totalPerPage: $request->get('per_page', 15),
+            filter: $request->filter
+        );
+
+        $filters = ['filter' => $request->get('filter', '')];
+
+        return view('dashboard.categorias.categorias_inativas', compact('categorias', 'filters'));
+    }
+
+    //=========================================================================================================
     public function delete(Request $request)
     {
+        if (Produto::where('categoria_id', $request->id)->first()) {
+            return redirect()->route('categorias.inativas')
+                ->with('error_disable', 'Impossível inativar a categoria, pois existem produtos que pertencem a ela.');
+        }
 
         $this->service->delete($request->id);
 
         return redirect()->route('show.categorias');
     }
 }
-
